requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
greenlet==3.0.0
idna==3.4
iniconfig==2.0.0
packaging==23.2
playwright==1.39.0
pluggy==1.3.0
pyee==11.0.1
pytest==7.4.3
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-slugify==8.0.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
urllib3==2.0.7
